+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 
 ## Unreleased
 
+- `allowedWikiLinks`: Support treating `index.md` as folder index route.
+
 
 ## 0.1.0.0